chore: Linting dnatracing; setup skeletons for remaining tests
+ Skeletons for remaining testing.
+ Pruning is moved to its own submodule to reduce the size of `skeletonize.py` which was failing `pylint`.
+ Linting docstrings of `dnatracing.py`.
