+++ tests/tracing/test_nodestats.py
@@ -1,0 +1,307 @@
+"""Test the nodestats module."""
+
+import pytest
+
+# from topostats.tracing.nodestats import nodeStats
+
+# pylint: disable=unnecessary-pass
+
+
+@pytest.mark.parametrize()
+def test_get_node_stats() -> None:
+    """Test of get_node_stats() method of nodeStats class."""
+    pass
+
+
+def test_check_node_errorless() -> None:
+    """Test of check_node_errorless() method of nodeStats class."""
+    pass
+
+
+def test_skeleton_image_to_graph() -> None:
+    """Test of skeleton_image_to_graph() method of nodeStats class."""
+    pass
+
+
+def test_graph_to_skeleton_image() -> None:
+    """Test of graph_to_skeleton_image() method of nodeStats class."""
+    pass
+
+
+def test_tidy_branches() -> None:
+    """Test of tidy_branches() method of nodeStats class."""
+    pass
+
+
+def test_keep_biggest_object() -> None:
+    """Test of keep_biggest_object() method of nodeStats class."""
+    pass
+
+
+def test_connect_close_nodes() -> None:
+    """Test of connect_close_nodes() method of nodeStats class."""
+    pass
+
+
+def test_highlight_node_centres() -> None:
+    """Test of highlight_node_centres() method of nodeStats class."""
+    pass
+
+
+def test_connect_extended_nodes() -> None:
+    """Test of connect_extended_nodes() method of nodeStats class."""
+    pass
+
+
+def test_connect_extended_nodes_nearest() -> None:
+    """Test of connect_extended_nodes_nearest() method of nodeStats class."""
+    pass
+
+
+#   def bounding_box
+#   def do_sets_touch() -> None:
+
+
+def test_find_branch_starts() -> None:
+    """Test of find_branch_starts() method of nodeStats class."""
+    pass
+
+
+def test_analyse_nodes() -> None:
+    """Test of analyse_nodes() method of nodeStats class."""
+    pass
+
+
+def test_sq() -> None:
+    """Test of sq() method of nodeStats class."""
+    pass
+
+
+def test_tri() -> None:
+    """Test of tri() method of nodeStats class."""
+    pass
+
+
+def test_auc() -> None:
+    """Test of auc() method of nodeStats class."""
+    pass
+
+
+def test_get_two_combinations() -> None:
+    """Test of get_two_combinations() method of nodeStats class."""
+    pass
+
+
+def test_cross_confidence() -> None:
+    """Test of cross_confidence() method of nodeStats class."""
+    pass
+
+
+def test_recip() -> None:
+    """Test of recip() method of nodeStats class."""
+    pass
+
+
+def test_per_diff() -> None:
+    """Test of per_diff() method of nodeStats class."""
+    pass
+
+
+def test_detect_ridges() -> None:
+    """Test of detect_ridges() method of nodeStats class."""
+    pass
+
+
+def test_get_box_lims() -> None:
+    """Test of get_box_lims() method of nodeStats class."""
+    pass
+
+
+def test_order_branch() -> None:
+    """Test of order_branch() method of nodeStats class."""
+    pass
+
+
+def test_order_branch_from_start() -> None:
+    """Test of order_branch_from_start() method of nodeStats class."""
+    pass
+
+
+def test_local_area_sum() -> None:
+    """Test of local_area_sum() method of nodeStats class."""
+    pass
+
+
+def test_get_vector() -> None:
+    """Test of get_vector() method of nodeStats class."""
+    pass
+
+
+def test_calc_angles() -> None:
+    """Test of calc_angles() method of nodeStats class."""
+    pass
+
+
+def test_pair_vectors() -> None:
+    """Test of pair_vectors() method of nodeStats class."""
+    pass
+
+
+def test_best_matches() -> None:
+    """Test of best_matches() method of nodeStats class."""
+    pass
+
+
+def test_create_weighted_graph() -> None:
+    """Test of create_weighted_graph() method of nodeStats class."""
+    pass
+
+
+def test_pair_angles() -> None:
+    """Test of pair_angles() method of nodeStats class."""
+    pass
+
+
+def test_gaussian() -> None:
+    """Test of gaussian() method of nodeStats class."""
+    pass
+
+
+def test_fwhm() -> None:
+    """Test of fwhm() method of nodeStats class."""
+    pass
+
+
+def test_fwhm2() -> None:
+    """Test of fwhm2() method of nodeStats class."""
+    pass
+
+
+def test_peak_height() -> None:
+    """Test of peak_height() method of nodeStats class."""
+    pass
+
+
+def test_lin_interp() -> None:
+    """Test of lin_interp() method of nodeStats class."""
+    pass
+
+
+def test_close_coords() -> None:
+    """Test of close_coords() method of nodeStats class."""
+    pass
+
+
+def test_order_branches() -> None:
+    """Test of order_branches() method of nodeStats class."""
+    pass
+
+
+def test_binary_line() -> None:
+    """Test of binary_line() method of nodeStats class."""
+    pass
+
+
+def test_coord_dist() -> None:
+    """Test of coord_dist() method of nodeStats class."""
+    pass
+
+
+def test_coord_dist_rad() -> None:
+    """Test of coord_dist_rad() method of nodeStats class."""
+    pass
+
+
+def test_above_below_value_idx() -> None:
+    """Test of above_below_value_idx() method of nodeStats class."""
+    pass
+
+
+def test_average_height_trace() -> None:
+    """Test of average_height_trace() method of nodeStats class."""
+    pass
+
+
+def test_fill_holes() -> None:
+    """Test of fill_holes() method of nodeStats class."""
+    pass
+
+
+def test_remove_re_entering_branches() -> None:
+    """Test of remove_re_entering_branches() method of nodeStats class."""
+    pass
+
+
+def test_only_centre_branches() -> None:
+    """Test of only_centre_branches() method of nodeStats class."""
+    pass
+
+
+def test_average_uniques() -> None:
+    """Test of average_uniques() method of nodeStats class."""
+    pass
+
+
+def test_compile_trace() -> None:
+    """Test of compile_trace() method of nodeStats class."""
+    pass
+
+
+def test_get_minus_img() -> None:
+    """Test of get_minus_img() method of nodeStats class."""
+    pass
+
+
+def test_remove_common_values() -> None:
+    """Test of remove_common_values() method of nodeStats class."""
+    pass
+
+
+def test_trace() -> None:
+    """Test of trace() method of nodeStats class."""
+    pass
+
+
+def test_reduce_rows() -> None:
+    """Test of reduce_rows() method of nodeStats class."""
+    pass
+
+
+def test_get_trace_segment() -> None:
+    """Test of get_trace_segment() method of nodeStats class."""
+    pass
+
+
+def test_comb_xyzs() -> None:
+    """Test of comb_xyzs() method of nodeStats class."""
+    pass
+
+
+def test_remove_duplicates() -> None:
+    """Test of remove_duplicates() method of nodeStats class."""
+    pass
+
+
+def test_order_from_end() -> None:
+    """Test of order_from_end() method of nodeStats class."""
+    pass
+
+
+def test_get_trace_idxs() -> None:
+    """Test of get_trace_idxs() method of nodeStats class."""
+    pass
+
+
+def test_get_visual_img() -> None:
+    """Test of get_visual_img() method of nodeStats class."""
+    pass
+
+
+def test_average_crossing_confs() -> None:
+    """Test of average_crossing_confs() method of nodeStats class."""
+    pass
+
+
+def test_minimum_crossing_confs() -> None:
+    """Test minimum_crossing_confs() method of nodeStats class."""
+    pass
