+++ tests/tracing/test_pruning.py
@@ -1,0 +1,97 @@
+"""Test the skeletonize module."""
+
+import pytest
+
+# from topostats.tracing.pruning import get_skeleton
+
+# pylint: disable=unnecessary-pass
+
+
+# Tests for pruneSkeleton class
+@pytest.mark.parametrize()
+def test_prune_skeleton() -> None:
+    """Test of prune_skeleton() method of pruneSkeleton class."""
+    pass
+
+
+def test_prune_method() -> None:
+    """Test of prune_method() method of pruneSkeleton class."""
+    pass
+
+
+def test_prune_topostats() -> None:
+    """Test of prune_topostats() method of pruneSkeleton class."""
+    pass
+
+
+def test_prune_conv() -> None:
+    """Test of prune_conv() method of pruneSkeleton class."""
+    pass
+
+
+# Tests for topostatsPrune class
+def test_topostats_prune_all_skeletons() -> None:
+    """Test of topostats_prune_all_skeletons() method of topostatsPrune class."""
+    pass
+
+
+def test_topostats_prune_by_length() -> None:
+    """Test of topostats_prune_by_length() method of topostatsPrune class."""
+    pass
+
+
+def test_topostats_find_branch_ends() -> None:
+    """Test of topostats_find_branch_ends() method of topostatsPrune class."""
+    pass
+
+
+# Tests for convPrune class
+def test_conv_prune_all_skeletons() -> None:
+    """Test of conv_prune_all_skeletons() method of convPrune class."""
+    pass
+
+
+def test_conv_prune_by_length() -> None:
+    """Test of conv_prune_by_length() method of convPrune class."""
+    pass
+
+
+# Tests for heightPruning class
+def test_conv_get_branch_mins() -> None:
+    """Test of conv_get_branch_mins() method of heightPruning class."""
+    pass
+
+
+def test_conv_get_branch_medians() -> None:
+    """Test of conv_get_branch_medians() method of heightPruning class."""
+    pass
+
+
+def test_conv_get_branch_middles() -> None:
+    """Test of conv_get_branch_middles() method of heightPruning class."""
+    pass
+
+
+def test_conv_get_abs_thresh_idx() -> None:
+    """Test of conv_get_abs_thresh_idx() method of heightPruning class."""
+    pass
+
+
+def test_conv_get_mean_abs_thresh_idx() -> None:
+    """Test of conv_get_mean_abs_thresh_idx() method of heightPruning class."""
+    pass
+
+
+def test_conv_get_iqr_thresh_idx() -> None:
+    """Test of conv_get_iqr_thresh_idx() method of heightPruning class."""
+    pass
+
+
+def test_conv_check_skeleton_one_object() -> None:
+    """Test of conv_check_skeleton_one_object() method of heightPruning class."""
+    pass
+
+
+def test_conv_remove_bridges() -> None:
+    """Test of remove_bridges() method of heightPruning class."""
+    pass
