+++ tests/tracing/test_skeletonize.py
@@ -541,3 +541,98 @@ def test_get_skeleton(  # pylint: disable=too-many-arguments
     assert skeleton.shape == shape
     assert skeleton.sum() == array_sum
     np.testing.assert_array_equal(skeleton, target)
+
+
+# pylint: disable=unnecessary-pass
+
+
+# Tests for topopstatsSkeletonize class
+def test_do_skeletonising_iteration() -> None:
+    """Test of do skeletonising iteration."""
+    pass
+
+
+def test_delete_pixel_subit1() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_delete_pixel_subit2() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_binary_thin_check_a() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_binary_thin_check_b_returncount() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_binary_thin_check_c() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_binary_thin_check_d() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_binary_thin_check_csharp() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_binary_thin_check_dsharp() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_final_skeletonisation_iteration() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_binary_final_thin_check_a() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_binary_final_thin_check_b() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_binary_thin_check_diag() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_get_local_pixels_binary() -> None:
+    """Test of  method."""
+    pass
+
+
+def test_do_skeletonising() -> None:
+    """Test of  method."""
+    pass
+
+
+# Miscellaneous functions
+def test_order_branch_from_start() -> None:
+    """Test of order_branch_from_start()."""
+    pass
+
+
+def test_rm_nibs() -> None:
+    """Test of rm_nibs()."""
+    pass
+
+
+def test_local_area_sum() -> None:
+    """Test local_area_sum()."""
+    pass
